[WIP] - put into methods
